fix(silver): use float64 for cd_municipio_ibge to match BQ FLOAT64 schema
Chunks with 100% municipality match produced numpy int64 (no NaN to force
float) which _normalize_for_bq_inplace did not convert, causing INT64/DOUBLE
schema mismatch on BQ load. Also harden normalizer to catch all numpy int kinds.

+++ dataops/depara_municipios.py
@@ -178,11 +178,11 @@ def join_tse_ibge(
         return df_tse
 
     df_ibge_clean = df_ibge.copy()
-    df_ibge_clean[ibge_key] = pd.to_numeric(df_ibge_clean[ibge_key], errors="coerce").astype(
-        "Int64"
-    )
+    df_ibge_clean[ibge_key] = pd.to_numeric(
+        df_ibge_clean[ibge_key], errors="coerce"
+    ).astype("float64")
     df_tse["cd_municipio_ibge"] = pd.to_numeric(
         df_tse["cd_municipio_ibge"], errors="coerce"
-    ).astype("Int64")
+    ).astype("float64")
 
     return df_tse.merge(df_ibge_clean, left_on="cd_municipio_ibge", right_on=ibge_key, how="left")
